+++ test/otpless_flutter_method_channel_test.dart
@@ -1,24 +1,13 @@
-import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:otpless_headless_flutter/otpless_flutter_method_channel.dart';
+import 'package:otpless_headless_flutter/otpless_flutter_platform_interface.dart';
 
 void main() {
-  MethodChannelOtplessFlutter platform = MethodChannelOtplessFlutter();
-  const MethodChannel channel = MethodChannel('otpless_headless_flutter');
-
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  setUp(() {
-    channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      return '42';
-    });
-  });
-
-  tearDown(() {
-    channel.setMockMethodCallHandler(null);
-  });
-
-  test('getPlatformVersion', () async {
-    expect(await platform.getPlatformVersion(), '42');
+  test('MethodChannelOtplessFlutter is a platform interface implementation',
+      () {
+    final MethodChannelOtplessFlutter platform = MethodChannelOtplessFlutter();
+    expect(platform, isA<OtplessFlutterPlatform>());
   });
 }