+++ test/otpless_flutter_test.dart
@@ -4,16 +4,18 @@ import 'package:otpless_headless_flutter/otpless_flutter_method_channel.dart';
 import 'package:otpless_headless_flutter/otpless_flutter_platform_interface.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-class MockOtplessFlutterPlatform 
+class MockOtplessFlutterPlatform
     with MockPlatformInterfaceMixin
     implements OtplessFlutterPlatform {
-
   @override
   Future<String?> getPlatformVersion() => Future.value('42');
 }
 
 void main() {
-  final OtplessFlutterPlatform initialPlatform = OtplessFlutterPlatform.instance;
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  final OtplessFlutterPlatform initialPlatform =
+      OtplessFlutterPlatform.instance;
 
   test('$MethodChannelOtplessFlutter is the default instance', () {
     expect(initialPlatform, isInstanceOf<MethodChannelOtplessFlutter>());
@@ -23,7 +25,7 @@ void main() {
     Otpless otplessHeadlessPlugin = Otpless();
     MockOtplessFlutterPlatform fakePlatform = MockOtplessFlutterPlatform();
     OtplessFlutterPlatform.instance = fakePlatform;
-  
+
     expect(await otplessHeadlessPlugin.getPlatformVersion(), '42');
   });
 }